+++ src/main/java/com/kyon/todolist/Controller/TodoListController.java
@@ -10,6 +10,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import java.sql.Date;
+import java.util.List;
 
 @Controller
 public class TodoListController {
@@ -20,14 +21,17 @@ public class TodoListController {
     public String todolist(Model model){
         Iterable<DbTodo> list = repository.findAll(Sort.by(Sort.Direction.ASC, "deadlineDt"));
         model.addAttribute("data", list);
+        model.addAttribute("size", ((List<DbTodo>) list).size());
         return "todolist";
     }
 
     @RequestMapping(value = "/new", method = RequestMethod.POST)
-    public String add(DbTodo data) {
-        data.setState(false);
-        data.setCreationDt(new Date(System.currentTimeMillis()));
-        this.repository.save(data);
+    public String add(DbTodo data, @RequestParam("comment") String comment) {
+        if(repository.findExactMatch(comment).size() == 0) {
+            data.setState(false);
+            data.setCreationDt(new Date(System.currentTimeMillis()));
+            this.repository.save(data);
+        }
         return "forward:/todolist";
     }
 
@@ -36,6 +40,21 @@ public class TodoListController {
         DbTodo data = this.repository.getOne(id);
         data.setState(true);
         this.repository.save(data);
+        return "forward:/todolist";
+    }
+
+    @RequestMapping(value = "/undone", method = RequestMethod.POST)
+    public String undone(@RequestParam("id") Integer id) {
+        DbTodo data = this.repository.getOne(id);
+        data.setState(false);
+        this.repository.save(data);
+        return "forward:/todolist";
+    }
+
+    @RequestMapping(value = "/delete", method = RequestMethod.POST)
+    public String delete(@RequestParam("id") Integer id) {
+        DbTodo data = this.repository.getOne(id);
+        this.repository.delete(data);
         return "forward:/todolist";
     }
 }
