+++ src/main/java/com/kyon/todolist/Database/DbRepository.java
@@ -10,4 +10,6 @@ import java.util.List;
 public interface DbRepository extends JpaRepository<DbTodo, Integer> {
     @Query("select a from DbTodo a where a.comment like %:keyword% order by a.deadlineDt asc")
     List<DbTodo> findComment(@Param("keyword") String keyword);
+    @Query("select a from DbTodo a where a.comment like :keyword")
+    List<DbTodo> findExactMatch(@Param("keyword") String keyword);
 }
