+++ src/main/resources/templates/edit.html
@@ -13,7 +13,7 @@
                 <td>
                     <form method="post" th:action="@{/comment}">
                         <input type="hidden" name="id" th:value="*{data.id}"/>
-                        <input type="text" name="comment" placeholder="todo"/>
+                        <input type="text" name="comment" placeholder="todo" maxlength="30" required/>
                         <input type="submit" value="Edit"/>
                     </form>
                 </td>
@@ -25,26 +25,8 @@
                 <td>
                     <form method="post" th:action="@{/deadlineDt}">
                         <input type="hidden" name="id" th:value="*{data.id}"/>
-                        <input type="date" name="deadlineDt" placeholder="date"/>
+                        <input type="date" name="deadlineDt" placeholder="date" required/>
                         <input type="submit" value="Edit"/>
-                    </form>
-                </td>
-            </tr>
-        </table>
-        <h2>Completed</h2>
-        <table>
-            <tr th:each="all:${all}"  th:if="*{all.state} eq true">
-                <td th:text="*{all.comment}"></td>
-                <td>
-                    <form method="post" th:action="@{/undone}">
-                        <input type="hidden" name="id" th:value="*{all.id}"/>
-                        <input type="submit" value="undone" class="undone"/>
-                    </form>
-                </td>
-                <td>
-                    <form method="post" th:action="@{/delete}">
-                        <input type="hidden" name="id" th:value="*{all.id}"/>
-                        <input type="submit" value="delete" class="done"/>
                     </form>
                 </td>
             </tr>
